working gimbal task
