+++ user/TASK/INS_task/INS_task.c
@@ -31,7 +31,6 @@
 #include "task.h"
 #include "USART_comms.h"
 #include <stdio.h>
-#include "test_task.h"
 
 #define IMUWarnBuzzerOn() buzzer_on(95, 10000) //开机陀螺仪校准蜂鸣器
