+++ user/TASK/gimbal_task/gimbal_task.c
@@ -13,51 +13,24 @@
 */
 
 #include "math.h"
+#include <stdio.h>
 #include "gimbal_task.h"
 #include "stm32f4xx.h"
-
+#include "USART_comms.h"
+#include "string.h"
 #include "FreeRTOSConfig.h"
 #include "FreeRTOS.h"
 #include "task.h"
 
-#define GIMBAL_TASK_INIT_TIME 201
-#define CONTROL_TIME 1
-
-#define RC_MIN -1.0
-#define RC_MAX 1.0
-
-#define ENCODER_MIN 0
-#define ENCODER_MAX 8191
-#define YAW_MIN 0
-#define YAW_MAX 8191
-#define PITCH_MIN 0
-#define PITCH_MAX 8191
-
-// These will later be produced from RC
-/* angle in radians*/
 float theta_setpoint;
 float phi_setpoint;
 float rc_channel_1;
 float rc_channel_2;
+Gimbal_t gimbal;
+Gimbal_Motor_t gimbal_pitch_motor;
 
-/*
-* Returns an angle between -pi and pi
-* Requires: an angle in radians
-* Returns: an equivalent angle in radians between -pi and pi
-*/
-float get_domain_angle(float alpha);
-
-/*
-* Gets smallest angle required to travel from alpha to beta
-* requires: two angles in radians
-* returns: angle in radians with magnitude < 2*pi
-*/
-float get_relative_angle(float alpha, float beta);
-
-/**
- * Maps float in a specified range to an int in a new range as a linear mapping function
- */
-int linear_map_float_to_int(float val, float val_min, float val_max, int out_min, int out_max);
+//UART mailbox
+char str[32] = {0};
 
 /**
   * @brief      Handles cases when RC joystick is not quite centered
@@ -75,23 +48,47 @@ int linear_map_float_to_int(float val, float val_min, float val_max, int out_min
     }
 
 
-void gimbalTask(void* parameters){
+void gimbal_task(void* parameters){
 
     vTaskDelay(GIMBAL_TASK_INIT_TIME);
 	
+    gimbal_pitch_motor.gimbal_motor_raw = get_Pitch_Gimbal_Motor_Measure_Point();
+    
+    fp32 pitch_signal;
+    int vision_signal;
+    
+    PidTypeDef pid;
+    fp32 pid_constants[3] = {pid_kp, pid_ki, pid_kd};
+    PID_Init(&pid, PID_POSITION, pid_constants, max_out, max_iout);
+    
 	while(1){	
         /* For now we assume channel 1 is left-right stick and channel 2 is dn-up stick*/
         /* For now using strictly encoder feedback for position */
-        theta_setpoint = linear_map_float_to_int(rc_channel_1, RC_MIN, RC_MAX, YAW_MIN, YAW_MAX);
-        phi_setpoint = linear_map_float_to_int(rc_channel_2, RC_MIN, RC_MAX, PITCH_MIN, PITCH_MAX);
-        
-        
+        // theta_setpoint = linear_map_float_to_int(rc_channel_1, RC_MIN, RC_MAX, YAW_MIN, YAW_MAX);
+        // phi_setpoint = linear_map_float_to_int(rc_channel_2, RC_MIN, RC_MAX, PITCH_MIN, PITCH_MAX);
 		// Gets update on position from encoders and gyro
 		// Comparison to setpoint
 		// run PID
-        vTaskDelay(CONTROL_TIME);
+        //vTaskDelay(CONTROL_TIME);
+        
+        // Get CAN received data
+        gimbal_pitch_motor.pos_raw = gimbal_pitch_motor.gimbal_motor_raw->ecd;
+        gimbal_pitch_motor.speed_raw = gimbal_pitch_motor.gimbal_motor_raw->speed_rpm;
+        gimbal_pitch_motor.current_raw = gimbal_pitch_motor.gimbal_motor_raw->given_current;
+
+        vision_signal = get_vision_signal();
+        
+        pitch_signal = PID_Calc(&pid, gimbal_pitch_motor.pos_raw, vision_signal);
+
+        // Turn gimbal motor
+        CAN_CMD_GIMBAL(0, pitch_signal, 0, 0);
+        
+        vTaskDelay(1);
+        
+        send_to_uart(gimbal_pitch_motor, pid, pitch_signal);  //Sending data via UART
 	}
 }
+
 
 int linear_map_float_to_int(float val, float val_max, float val_min, int out_min, int out_max){
     float percentage = val / (val_max - val_min);
@@ -113,10 +110,60 @@ float get_domain_angle(float alpha){
 			done = TRUE;
 		}
 	}
+    return 0.0;
 }
 
 
 float get_relative_angle(float alpha, float beta){
 	return get_domain_angle(beta - alpha);
 }
-	
+
+
+/** 
+ * @brief  Reads vision instruction from UART and cap to certin values
+ * @param  None
+ * @retval Vision signal in range of 0 and 8191
+ */
+int get_vision_signal(void) {
+    int vision_signal = 5700;  // TODO: Get real values from vision
+        
+    while (vision_signal > 8191) {
+        vision_signal -= 8191;
+    }
+    while (vision_signal < 0) {
+        vision_signal+= 8192;
+    }
+    return vision_signal;
+}
+
+
+/** 
+ * @brief  
+ * @param  
+ * @retval None
+ */
+void send_to_uart(Gimbal_Motor_t gimbal_yaw_motor, PidTypeDef pid, fp32 pitch_signal) 	
+{
+    char str[20]; //uart data buffer
+
+    sprintf(str, "position: %d\n\r", gimbal_yaw_motor.pos_raw);
+    serial_send_string(str);
+
+    sprintf(str, "speed: %d\n\r", gimbal_yaw_motor.speed_raw);
+    serial_send_string(str);       
+
+    sprintf(str, "current: %d\n\r", gimbal_yaw_motor.current_raw);
+    serial_send_string(str);
+
+    sprintf(str, "motor kp: %f\n\r", pid.Kp);
+    serial_send_string(str);
+
+    sprintf(str, "motor kd: %f\n\r", pid.Kd);
+    serial_send_string(str);    
+
+    sprintf(str, "motor ki: %f\n\r", pid.Ki);
+    serial_send_string(str);
+    
+    sprintf(str, "pitch signal: %f\n\r", pitch_signal);
+    serial_send_string(str);
+}
