+++ user/TASK/start_task/start_task.c
@@ -4,26 +4,31 @@
 #include "FreeRTOS.h"
 #include "task.h"
 
-#include "test_task.h"
 #include "revolver_task.h"
 #include "INS_task.h"
 #include "chassis_task.h"
+#include "gimbal_task.h"
+#include "test_task.h"
 
 #define START_TASK_PRIO 1
 #define START_STK_SIZE 512
 static TaskHandle_t start_task_handler;
 
-#define TEST_TASK_PRIO 4
-#define TEST_STK_SIZE 256
-static TaskHandle_t test_task_handler;
-
 #define INS_TASK_PRIO 2
 #define INS_STK_SIZE 512
 static TaskHandle_t chassis_task_handler;
 
-#define CHASSIS_TASK_PRIO 2
+#define CHASSIS_TASK_PRIO 3
 #define CHASSIS_STK_SIZE 512
 static TaskHandle_t INS_task_handler;
+
+#define GIMBAL_TASK_PRIO 4
+#define GIMBAL_STK_SIZE 512
+static TaskHandle_t gimbal_task_handler;
+
+#define TEST_TASK_PRIO 5
+#define TEST_STK_SIZE 512
+static TaskHandle_t test_task_handler;
 
 #define REV_TASK_PRIO 10
 #define REV_STK_SIZE 256
@@ -32,15 +37,7 @@ static TaskHandle_t revolver_task_Handler;
 void start_task(void *pvParameters)
 {
     taskENTER_CRITICAL();
-
-    xTaskCreate((TaskFunction_t) test_task,
-            (const char *)"test_task",
-            (uint16_t) TEST_STK_SIZE,
-            (void *)NULL,
-            (UBaseType_t)TEST_TASK_PRIO,
-            (TaskHandle_t *)&test_task_handler);
-
-                                                            
+                                                        
     xTaskCreate((TaskFunction_t) INS_task,
             (const char *)"INS_task",
             (uint16_t) INS_STK_SIZE,
@@ -54,8 +51,21 @@ void start_task(void *pvParameters)
             (void *)NULL,
             (UBaseType_t)CHASSIS_TASK_PRIO,
             (TaskHandle_t *)&chassis_task_handler);
-                            
 						
+    xTaskCreate((TaskFunction_t) gimbal_task,
+            (const char *)"gimbal_task",
+            (uint16_t) GIMBAL_STK_SIZE,
+            (void *)NULL,
+            (UBaseType_t)GIMBAL_TASK_PRIO,
+            (TaskHandle_t *)&gimbal_task_handler);
+
+    xTaskCreate((TaskFunction_t) test_task,
+            (const char *)"test_task",
+            (uint16_t) TEST_STK_SIZE,
+            (void *)NULL,
+            (UBaseType_t)TEST_TASK_PRIO,
+            (TaskHandle_t *)&test_task_handler);
+            
     xTaskCreate((TaskFunction_t) revolver_task,
             (const char *)"revolver_task",
             (uint16_t) REV_STK_SIZE,
