+++ user/TASK/test_task/test_task.h
@@ -12,40 +12,17 @@
 
 #include "CAN_receive.h"
 #include "main.h"
-#include "pid.h"
 
-
-/******************** PID Constants (May need to rename...) ********************/
-
-#define pid_kp 40.0f
-#define pid_ki 0.0f
-#define pid_kd 0.0f
-#define max_out 5000.0f
-#define max_iout 0
-
-
-/******************** Public Definitions & Structs ********************/
-
-typedef struct 
-{
-	const motor_measure_t *gimbal_motor_raw;
-	uint16_t pos_raw;
-	uint16_t speed_raw;
-	uint16_t current_raw;
-} Gimbal_Motor_t;
-
-typedef struct 
-{
-	Gimbal_Motor_t *yaw_motor;
-	Gimbal_Motor_t *pitch_motor;
-    const fp32 *angle_reading_raw;
-	const fp32 *gyro_reading_raw;
-	const fp32 *acce_reading_raw;
-} Gimbal_t;
+//Reading angle, gyro, and accelerometer data and printing to serial
+static void test_imu_readings(uint8_t angle, uint8_t gyro, uint8_t acce);
+//Turns a gimbal motor (GM6020) and outputs its position, rpm, and current
+static void test_GM6020(void);
+//Enables Debug of P19 (chassis) motor
+static void test_P19(int id);
 
 
 /******************** Task/Functions Called Outside ********************/
 //Task used for testing. Usually left blank.
 void test_task(void *pvParameters);
-	
+
 #endif
